requirements.txt
Django==2.0
pycryptodome==3.6.6
djangorestframework==3.7.7
django-cors-headers==2.1.0
requests==2.18.4
django-filter==1.1.0
psycopg2==2.7.4
django-rest-swagger==2.2.0
numpy==1.15.4
sympy==1.3
django-heroku
gunicorn
pycryptodome==3.6.6
django-axes==4.4.3
django-bootstrap4==0.0.7
